+++ pages/[lang]/misc/index.ts
@@ -1,0 +1,4 @@
+import {MiscPostList} from '../../../src/components/pages/posts/misc/list/list';
+
+
+export default MiscPostList;
