+++ pages/[lang]/quest/index.ts
@@ -1,0 +1,4 @@
+import {QuestPostList} from '../../../src/components/pages/posts/quest/list/list';
+
+
+export default QuestPostList;
