+++ src/components/pages/posts/misc/list/list.tsx
@@ -1,22 +1,35 @@
 import React from 'react';
 
-import {MiscPostListEntry} from '../../../../../api-def/api';
-import {LinkGenerator} from '../../../../elements/posts/list/entry';
+import {GeneralPath, PostPath} from '../../../../../const/path/definitions';
+import {useI18n} from '../../../../../i18n/hook';
+import {makePostUrl} from '../../../../../utils/path/make';
+import {ApiRequestSender} from '../../../../../utils/services/api/requestSender';
 import {PostList} from '../../../../elements/posts/list/list';
+import {PostListPage} from '../../../../elements/posts/list/page';
 import {MiscEntryBadge} from './listBadges';
 
 
-type Props = {
-  entries: Array<MiscPostListEntry>,
-  generateLink: LinkGenerator,
-};
+export const MiscPostList = () => {
+  const {t, lang} = useI18n();
 
-export const MiscPostList = ({entries, generateLink}: Props) => {
+  const title = t((t) => t.meta.inUse.post.misc.list.title);
+
   return (
-    <PostList
-      entries={entries}
-      generateLink={generateLink}
-      renderPostBadge={(props) => <MiscEntryBadge {...props}/>}
+    <PostListPage
+      title={title}
+      fnFetchList={ApiRequestSender.miscList}
+      postManageBarProps={{
+        newButtons: [{pathname: GeneralPath.MISC_NEW}],
+      }}
+      renderPostEntries={(response) => (
+        <PostList
+          entries={response.posts}
+          generateLink={(postId) => (
+            makePostUrl(PostPath.MISC, {pid: postId, lang})
+          )}
+          renderPostBadge={(props) => <MiscEntryBadge {...props}/>}
+        />
+      )}
     />
   );
 };
